Created a functioning auth server, currently allows all requests
I removed the web token library as I felt it was rediculious to have
a fairly large library for such small usage. Also, I couldn't
figure out how to manually edit the header to add the kid field.

Likewise, I removed the UUID library and implemented the needed code
to generate a v4 UUID.

Docker is happy with the responses, however the server doesn't
actually do any authentication at the moment. It will allow
whatever the client asks for. It doesn't even check for passwords.